requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Caching==2.1.0
Flask-Mail==0.9.1
Flask-Login==0.6.3
Werkzeug==2.3.7
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
Jinja2==3.1.3
MarkupSafe==2.1.3
itsdangerous==2.1.2
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
asyncio==3.4.3
razorpay==1.4.1
pytz==2023.3
Pillow==11.2.1
python-dateutil==2.8.2
uuid==1.30
scrapy==2.11.0
reportlab==4.0.8
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==41.0.7
marshmallow==3.20.1

# For processing and analysis
numpy==2.2.1
pandas>=2.2.0
Markdown==3.5.1

# For HTML parsing and web scraping
lxml==4.9.3
urllib3==2.0.7
parse==1.20.0

# For testing
pytest==7.4.3

# For development
black==23.11.0
flake8==6.1.0